recoveryTools/requirements.txt
Pillow>=9.0.0
imagehash>=4.3.1
numpy>=1.20.0
ffmpeg-python>=0.2.0